+++ packages/atomic/cypress/integration/result-list/result-components/result-link.cypress.ts
@@ -9,6 +9,7 @@ import {
   resultLinkComponent,
   ResultLinkSelectors,
 } from './result-link-selectors';
+import * as CommonAssertions from '../../common-assertions';
 
 interface ResultLinkProps {
   target?: '_self' | '_blank' | '_parent' | '_top';