+++ packages/atomic/src/components/result-template-components/atomic-result-badge/atomic-result-badge.tsx
@@ -53,7 +53,7 @@ export class AtomicResultBadge {
         {this.field ? (
           <atomic-result-text field={this.field}></atomic-result-text>
         ) : (
-          this.label
+          <atomic-text value={this.label ?? ''}></atomic-text>
         )}
       </span>
     );